requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.1.0
faster-whisper>=1.1.0
elevenlabs>=1.50.0
